requirements.txt
requests
tweepy
textblob
numpy
pillow